subprocess.check_call is not available on python 2.4
alas, python 2.4 is used on ent1

git-svn-id: http://subversion.cru.fr/esup-tools/PAGScreator/trunk@61 93fd7c73-ba33-0410-a30b-86d151360d8f

+++ PAGScreator.py
@@ -88,7 +88,8 @@ def write_to_file(doc, name):
 def checkAndIndent(xmlFile):
     tmp_file = xmlFile + ".tmp"
     os.rename(xmlFile, tmp_file)
-    subprocess.check_call(["xmllint", "--format", "-o", xmlFile, tmp_file])
+    if subprocess.call(["xmllint", "--format", "-o", xmlFile, tmp_file]) != 0:
+        exit("xmllint failed. install libxml2-utils?")
     os.remove(tmp_file)
 
 def configureLogger(logger):
